+++ pkg/xrr/xrrtest/all_test.go
@@ -17,19 +17,19 @@ func TstError() error {
 	e := xrr.New("msg", "EC", m)
 
 	m = xrr.Meta().Int("A", 2).Int("int", 2).Option()
-	e = xrr.Wrap[edXrrTest](e, m)
+	e = xrr.WrapUsing[edXrrTest](e, m)
 
 	m = xrr.Meta().Int("A", 3).Int64("int64", 3).Option()
-	e = xrr.Wrap[edXrrTest](e, m)
+	e = xrr.WrapUsing[edXrrTest](e, m)
 
 	m = xrr.Meta().Int("A", 4).Float64("float64", 4).Option()
-	e = xrr.Wrap[edXrrTest](e, m)
+	e = xrr.WrapUsing[edXrrTest](e, m)
 
 	m = xrr.Meta().Int("A", 5).Bool("bool", true).Option()
-	e = xrr.Wrap[edXrrTest](e, m)
+	e = xrr.WrapUsing[edXrrTest](e, m)
 
 	m = xrr.Meta().Int("A", 6).Time("tim", tim).Option()
-	e = xrr.Wrap[edXrrTest](e, m)
+	e = xrr.WrapUsing[edXrrTest](e, m)
 
 	return e
 }